pyproject.toml
[build-system]
build-backend = 'hatchling.build'
requires = ['hatchling']


[project]
name = 'alphagenome_research'
description = 'AlphaGenome Research.'
readme = 'README.md'
dynamic = ['version']
license = { file = 'LICENSE' }
requires-python = '>=3.11'
authors = [
    {name = 'Google LLC'},
    {email = 'alphagenome@google.com'},
]
keywords = [
    'python',
    'machine learning',
    'genomics'
]
classifiers=[
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]
dependencies=[
    # keep-sorted start
    'absl-py',
    'alphagenome',
    'anndata',
    'chex',
    'dm-haiku',
    'einshape',
    'etils[epath]',
    'huggingface_hub',
    'jax',
    'jaxtyping',
    'kagglehub',
    'numpy',
    'optax',
    'orbax',
    'pandas',
    'pyBigWig',
    'pyarrow',
    'pyfaidx',
    'pyranges',
    'tensorflow',
    # keep-sorted end
]

[project.urls]
Repository = 'https://github.com/google-deepmind/alphagenome_research'
Documentation = 'https://www.alphagenomedocs.com/'

[project.optional-dependencies]
dev = [ 'hatch' ]

[tool.hatch.version]
path = 'src/alphagenome_research/__init__.py'

[tool.hatch.envs.default]
installer = 'uv'

[tool.hatch.envs.hatch-test]
default-args = []
extra-dependencies=['typeguard==2.13.3']
parallel = true

[tool.hatch.envs.hatch-test.env-vars]
MPLBACKEND = 'agg'

[[tool.hatch.envs.hatch-test.matrix]]
# Use hatch test --all to run tests on all supported Python versions.
python = ['3.13', '3.12', '3.11']

[tool.hatch.envs.check]
dependencies = [
  'nbconvert',
  'pyink>=24.3.0',
  'pylint>=2.6.0',
]
# Do not install dependencies for the check environment.
detached = true

[tool.hatch.envs.check.scripts]
format = 'pyink . --check --diff'
lint = 'pylint .'
notebook_lint = 'jupyter nbconvert colabs/*.ipynb --to html'
all = [
  'format',
  'lint',
  'notebook_lint',
]

[tool.pyink]
# Formatting configuration to follow Google style-guide
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true
